fix: harden bridge receivers per adversarial review (6 findings)
Pre-push multi-agent review of the branch diff confirmed 6 defects;
all fixed with regression tests (47 passing):

1. Out-of-order retries could resurrect removed deals/businesses (the
   outbox re-signs retries with fresh timestamps, so the replay window
   gives no ordering). Now the sender's payload updated_at is persisted
   (deals.source_updated_at, businesses.bridge_updated_at) and a stale
   event's field changes are skipped — still recorded + 200'd.
2. discount_size was typed int but the frozen contract says number
   (sender zod allows 12.5) — fractional-discount deals would 422 and
   dead-letter forever. Now float end-to-end.
3. Non-ASCII X-HC-Signature header made compare_digest raise TypeError
   → unauthenticated 500s. Now compared as bytes → clean 401.
4. Any IntegrityError was treated as an eventId duplicate and 200'd
   without a durable commit. Now the ledger is re-checked: genuine
   replay → 200 duplicate; anything else → 500 so the sender retries.
5. Request body was buffered unbounded pre-auth and stored in
   bridge_events.payload. Now capped at 64 KB (413 past that).
6. is_active != False hid legacy NULL rows (SQL three-valued logic,
   opposite of the comment's claim). Now IS NOT false — NULL-safe.

+++ backend/schemas.py
@@ -105,7 +105,7 @@ class HybridCardDealPayload(BaseModel):
     sub_type: Optional[str] = None
     title: str
     short_description: Optional[str] = None
-    discount_size: int = 0  # marker sizing ONLY — stored, never ranked on (contract §7)
+    discount_size: float = 0  # contract type is number (12.5 is valid) — stored, never ranked on (§7)
     lat: float
     lng: float
     hours: Optional[str] = None
